+++ app/src/main/java/cc/scottland/sketchpad/utils/Utils.java
@@ -8,17 +8,21 @@ import cc.scottland.sketchpad.shapes.Point;
 
 public class Utils {
 
-    public static int distance(Point p1, Point p2) {
-        int dx = p1.x - p2.x;
-        int dy = p1.y - p2.y;
-        return (int) (Math.sqrt(dx * dx + dy * dy));
+    public static float distance(Point p1, Point p2) {
+
+        float dx = p1.x - p2.x;
+        float dy = p1.y - p2.y;
+
+        return (float)Math.sqrt(dx * dx + dy * dy);
     }
 
     public static double dot(Point p1, Point p2) {
+
         return p1.x * p2.x + p1.y * p2.y;
     }
 
     public static double angle(Point p1, Point p2) {
+
         return Math.toDegrees(Math.atan2(p2.y - p1.y, p2.x - p1.x));
     }
 }
